Move test_unregister_filter to test_h5z module
