+++ h5py/tests/test_filters.py
@@ -14,9 +14,8 @@
 import os
 import numpy as np
 import h5py
-import pytest
 
-from .common import ut, TestCase, insubprocess
+from .common import ut, TestCase
 
 
 class TestFilters(TestCase):
@@ -85,14 +84,6 @@ def test_filter_ref_obj_eq():
 
     assert gzip8 == h5py.filters.Gzip(level=8)
     assert gzip8 != h5py.filters.Gzip(level=7)
-
-
-@pytest.mark.mpi_skip
-@insubprocess
-def test_unregister_filter(request):
-    if h5py.h5z.filter_avail(h5py.h5z.FILTER_LZF):
-        res = h5py.h5z.unregister_filter(h5py.h5z.FILTER_LZF)
-        assert res
 
 
 @ut.skipIf(not os.getenv('H5PY_TEST_CHECK_FILTERS'),  "H5PY_TEST_CHECK_FILTERS not set")
